backend/requirements.txt
absl-py==2.1.0
accelerate==1.12.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.16.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==4.12.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.2
captum==0.7.0
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cryptography==46.0.3
cv==1.0.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.121.2
fastjsonschema==2.21.1
fastuuid==0.14.0
ffmpy==1.0.0
filelock==3.16.1
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.12.0
ftfy==6.3.1
fuzzywuzzy==0.18.0
gast==0.6.0
gensim==4.3.3
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.56.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
gradio==5.49.1
gradio_client==1.13.3
graphviz==0.21
greenlet==3.0.3
groovy==0.1.2
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
hf-xet==1.2.0
httpcore==1.0.5
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.4.1
idna==3.6
imbalanced-learn==0.12.4
imblearn==0.0
importlib-metadata==7.0.1
importlib_resources==6.5.2
imutils==0.5.4
ipykernel==6.29.4
ipython==8.24.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.5
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.83
langchain-huggingface==0.3.1
langchain-ollama==0.3.6
langchain-text-splitters==0.3.11
langsmith==0.6.8
lazy_loader==0.4
Levenshtein==0.27.1
libclang==18.1.1
librosa==0.10.2.post1
lightgbm==4.5.0
litellm==1.80.10
llama_cpp_python==0.3.4
llvmlite==0.43.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mc_bin_client==1.0.1
mdurl==0.1.2
mistune==3.1.1
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.25.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauthlib==3.3.1
ollama==0.5.1
open_clip_torch==3.2.0
openai==2.13.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.14.0
optuna==4.4.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.3
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.1
plotly==6.0.0
pooch==1.8.2
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycocoevalcap==1.2
pycocotools==2.0.11
pycparser==2.22
pydantic==2.11.0
pydantic-settings==2.7.0
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.25.5
pymupdf4llm==0.0.22
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==5.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2024.2
pywin32==311
pywinpty==3.0.3
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
ruff==0.14.5
safehttpx==0.1.7
safetensors==0.5.2
scikit-fuzzy==0.5.0
scikit-learn==1.3.0
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.3.1
shap==0.46.0
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.49.3
sympy==1.13.1
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.18.0
threadpoolctl==3.5.0
tiktoken==0.12.0
timm==1.0.22
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
transformers==5.1.0
typed-argument-parser==1.10.1
typer==0.20.0
typer-slim==0.20.0
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.1
uuid_utils==0.14.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
win-inet-pton==1.1.0
wrapt==1.17.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.18.1
zstandard==0.25.0
